refactor: view 통합
